Created seperate files for API calls for each table to make it more readable and easier to find
